spec 039: catalog maintenance - the check that finds what drifted, the repair that mends on purpose (#120)
* spec 039: draft - maintenance, telling an admin what the catalog can no longer do

A virtual catalog describes objects in sources it does not own, and nothing asks
whether its claims still hold. Probed: an object that renames its columns, then
the source drops one - the principal's ordinary SELECT * answers

  Binder Error: Referenced column "ssn_raw" not found in FROM clause!
  Candidate bindings: "internal_tenant"

Two physical names, to a principal who was only ever shown id, tenant, ssn. And
while the read fails outright, duckdb_columns() narrows to id, tenant, so the
metadata describes an object no query can return.

The draft settles the part that is a decision rather than a feature: repair is not
automatic, because the obvious self-heal - drop the vanished column from the
object's list - also drops any mask on it, which is the "silently skipped
protection" spec 038 refuses a mask over. So the surface is a diagnostic
(acl_check_catalog) plus an explicit repair that names what it gives up, and the
listing tells the truth about a broken object rather than quietly describing a
narrower one.

Not implemented; the backlog links here.


* spec 039: catalog maintenance - the check that finds what drifted, the repair that mends on purpose

acl_check_catalog([vcat]) / CHECK VIRTUAL CATALOG c: a table function, the operator's (denied to a
principal; catalog-scoped under a role's manage), that probes every stored fact of a catalog against
the source OFF the query path - spec 065's "no probing, anywhere" holds for the resolver - and
answers one row per finding with the repair to paste: source_missing, column_missing,
definition_broken, schema_stale, rls_broken / rls_unchecked, grant_column_missing, mask_broken,
schema_missing, expansion_stale, reference_dangling. It writes nothing.

acl_repair_relation(vcat, vname, action[, spec]) / REPAIR VIRTUAL TABLE c.n REMAP (v = expr, ...) |
DROP MISSING COLUMNS [AND MASKS]: mends a declared COLUMNS list on purpose. remap probes each new
expression to bind before anything is written; drop_missing removes the entries that cannot bind
and REFUSES while an object grant masks one of those names - spec 038's rule, a protection is never
dropped silently; AND MASKS removes those mask items by name and counts them; a catalog grant's mask
is never touched (it protects the column elsewhere, and the check reports the object's refusal as
mask_broken). The grants' projections and verdicts are re-probed after.

The tables listings mark a broken declared-list object in its comment (acl: broken - declared
column(s) ... no longer exist in the source) by a join over catalog facts and information_schema,
never a probe; the columns surface keeps the contract as written. The management batch learns the
one compiled form that is a table function in FROM (MgmtCallName, ProvenanceOf).

Tests: test/sql/acl_catalog_maintenance.test (175 assertions). Whole test/sql (60 files), test-cpp,
harness, test-flight, door e2e (postgres/ducklake/pair) and integration green locally. Docs:
management-sql (a section + the reference rows), security (the compensating control exists),
CLAUDE.md; both backlogs updated; the spec is implemented.


* spec 039 review: ANALYZE re-derives a table's projection, a check judged live, tidy across the tree

Two adversarial passes over spec 039, applied:
- schema_stale on a declared-list table named ANALYZE VIRTUAL TABLE, which re-derived only views
  and macros: CatalogRefreshSchema now re-derives a table's projection schema too - and carries
  the declared nullability marks and column comments through every re-derivation (it used to
  drop them for views and macros);
- a grant's items were judged against a view's STORED schema - exactly what may be stale - so a
  catalog-wide bare name that matched nothing any more went unreported; judged against the declared
  list, else what the source binds to now;
- the listing mark called a bare-word function (`who = current_user`) a missing column while the
  check was clean: SQL's bare-word functions and constants are not judged; and the mark is one
  grouped scan of information_schema.columns per listing instead of a correlated subquery per
  relation on every principal's listing;
- DROP MISSING COLUMNS AND MASKS is several catalog writes, so the order is what keeps a failure
  midway harmless: grants lose their mask items first (fewer items read less, never more), the
  object's list is cut after;
- the spec's vocabulary, refusal text and repair column now say what the code does (no
  mask_orphaned; a catalog grant's mask is never touched; object grants are re-granted through
  acl_grant_object, which has no ALTER form); duckdb_views() never carries the mark (a view has no
  declared list).

Tests: every problem of the vocabulary is produced at least once (rls_unchecked, schema_stale on a
table and on a view, mask_broken from an object grant, grant_column_missing for a catalog grant with
object `*`, expansion_stale in both halves with REFRESH PRUNE, reference_dangling for an end and for
a column, definition_broken for a macro template, a scalar alias and a view, source_missing on a
bare alias and a declared table), the binder error is pinned, the no-argument check walks a second
catalog with a broken object of its own, a computed column written as a bare-word function is not
marked. 233 assertions.

clang-tidy across our sources (`make tidy`, new: the repo's duckdb .clang-tidy over the compile
database, with the SDK sysroot macOS needs): by-value shared_ptr / function parameters made const
references, the deprecated DataChunk::SetValue/SetCardinality replaced, a use-after-move in the
rewriter's insert guard restructured (the guard now rides the first column outside the loop), an
int-typed year constant, `long long` casts dropped, a std::move into a const reference dropped, the
two deliberate process-wide caches and the thread-local seam marked NOLINT with their reasons.
Unused includes dropped where clang-include-cleaner and the compiler agree (`duckdb.hpp` and
`<arrow/table.h>` were transitively load-bearing and stay).

Whole test/sql (60 files), test-cpp, harness, test-flight, door e2e (postgres/ducklake/pair) and
integration green locally; `make tidy` clean.


---------

+++ src/acl_policy_catalog.cpp
@@ -1570,8 +1570,7 @@ string PolicyStore::ResolveIssuerKeys(const IssuerConfig &config, const string &
 		NoteDenyReason(Reason::SOURCE_ERROR); // the same cause as above: the keys' source, not the principal
 		throw BinderException("acl_rewrite: token rejected: the keys of issuer \"%s\" were last read %lld seconds "
 		                      "ago and \"%s\" is still unreadable (%s); acl_jwks_max_stale is %lld",
-		                      config.issuer, static_cast<long long>(now - entry.fetched_at), config.jwks_uri,
-		                      entry.error, static_cast<long long>(max_stale));
+		                      config.issuer, now - entry.fetched_at, config.jwks_uri, entry.error, max_stale);
 	}
 	return entry.keys_json;
 }
